requirements.txt
openai==1.14.0
python-dotenv==1.0.1
colorama==0.4.6
